reset logo form Fixes #71
