+++ html/index.html
@@ -404,6 +404,7 @@
           <div id="qrstylelogopreview" class="placeholder"></div>
           <input type="hidden" name="qrstylelogourl" id="qrstylelogourl" />
           <button name="qrstylelogo" id="qrstylelogo">Upload Logo</button>
+          <button name="qrstylelogoremove" id="qrstylelogoremove">Remove Logo</button>
         </details>
         <details>
           <summary>
